move migrations,models, and seeders into db folder
